Fixed Linux warnings

+++ ApplicationCode/ModelVisualization/RivReservoirPipesPartMgr.cpp
@@ -144,7 +144,7 @@ void RivReservoirPipesPartMgr::updatePipeResultColor(size_t frameIndex)
 //--------------------------------------------------------------------------------------------------
 const std::vector< std::vector <cvf::Vec3d> >* RivReservoirPipesPartMgr::centerLineOfWellBranches(int wellIdx)
 {
-    if (wellIdx < m_wellPipesPartMgrs.size())
+    if (wellIdx < static_cast<int>(m_wellPipesPartMgrs.size()))
     {
         return &(m_wellPipesPartMgrs[wellIdx]->centerLineOfWellBranches());
     }
